Specify CareLink server from command line
Added ability to specify which CareLink server to use from the command line.  This is useful if you aren't in the EU region as the request would succeed but return a binary blob that `javaobj` couldn't decode causing the program to fail with less than helpful diagnostics.  The program still defaults to the EU server.  `io.StringIO` has been replaced with `io.BytesIO` which removes the need to cast to a `unicode` object.